Replace Current Date with Clock Time number node
Clock Time outputs seconds since midnight as a Float, matching the
pattern of System Time and Patch Time. Date Formatter now accepts
a Float input (seconds) instead of Date, so it pairs naturally with
Clock Time via a connection.
