+++ Fabric/Nodes/NodeRegistry.swift
@@ -281,6 +281,7 @@ public class NodeRegistry {
         // Number
         CurrentTimeNode.self,
         SystemTimeNode.self,
+        ClockTimeNode.self,
         NumberNode.self,
         TimelineNode.self,
         NumberUnaryOperator.self,
@@ -325,7 +326,6 @@ public class NodeRegistry {
         StringFormatterNode.self,
         StringSeparatorNode.self,
         DateFormatterNode.self,
-        CurrentDateNode.self,
         DirectoryScannerNode.self,
         StringToTimecodeNode.self,
 //        ConvertToStringNode.self,
